pyproject.toml
[project]
name = "todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.120.0",
    "sqlalchemy>=2.0.44",
    "pydantic-settings>=2.11.0",
    "python-jose>=3.5.0",
    "uvicorn",
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "anyio>=4.11.0",
    "pytest-asyncio>=1.3.0",
    "pytest-tornasync>=0.6.0.post2",
    "pytest-trio>=0.8.0",
    "pytest-twisted>=1.14.3",
    "twisted>=25.5.0",
    "tzlocal>=5.3.1",
    "testcontainers[postgresql]>=4.14.1",
    "asyncpg>=0.31.0",
    "pyjwt>=2.12.1",
]

[tool.pytest.ini_options]
pythonpath = [".", "src"]
asyncio_mode = "auto"


